+++ src/main/java/com/codetaylor/mc/dropt/modules/dropt/rule/data/RuleDrop.java
@@ -4,5 +4,6 @@ public class RuleDrop {
 
   public RuleDropSelector selector = new RuleDropSelector();
   public RuleDropItem item = new RuleDropItem();
+  public RandomFortuneInt xp = new RandomFortuneInt(0, 0);
 
 }